+++ src/Apiresponse.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Yormy\Apiresponse;
-
-class Apiresponse
-{
-}